src/requirements/common.txt
Django>=1.8,<1.8.99
django-annoying==0.8.4
django-hijack==2.1.3
django-hijack-admin==2.1.3
django-markdown-deux==1.0.5
django-paypal==0.2.7
django-pure-pagination==0.2.1
django-simple-captcha==0.5.3
django-webpack-loader==0.4.1
feedparser==5.2.1
google-api-python-client==1.6.2
markdown==2.6.8
MySQL-python==1.2.5
Pillow==4.0.0
postmarkup==1.2.2
python-dateutil==2.6.0
pytils==0.3
